refactor: move game-manager into separate package
