+++ src/battleship_pygame_lan/game_manager/__init__.py
@@ -1,0 +1,4 @@
+from .enums import GuiEvent
+from .game_manager import GameManager
+
+__all__ = ["GameManager", "GuiEvent"]
