+++ src/battleship_pygame_lan/game_manager/enums.py
@@ -1,0 +1,13 @@
+from enum import Enum, auto
+
+
+class GuiEvent(Enum):
+    """
+    Enum representing type of action gui should take.
+    For example: make some kind of sound, show some text etc.
+    """
+
+    ShotMade = auto()
+    ShotHit = auto()
+    ShotMissed = auto()
+    ShotMarked = auto()
