+++ src/battleship_pygame_lan/game_manager/game_manager.py
@@ -1,5 +1,4 @@
 import socket
-from enum import Enum, auto
 from logging import getLogger
 from queue import Empty, Queue
 
@@ -12,19 +11,9 @@ from battleship_pygame_lan.logic import (
 )
 from battleship_pygame_lan.network import GameState, NetworkClient, PayloadTypes
 
+from .enums import GuiEvent
+
 logger = getLogger(__name__)
-
-
-class GuiEvent(Enum):
-    """
-    Enum representing type of action gui should take.
-    For example: make some kind of sound, show some text etc.
-    """
-
-    ShotMade = auto()
-    ShotHit = auto()
-    ShotMissed = auto()
-    ShotMarked = auto()
 
 
 class GameManager:
